增加升级订单类型

+++ shop-web/src/main/java/com/masiis/shop/web/platform/service/user/ComUserAccountService.java
@@ -90,7 +90,7 @@ public class ComUserAccountService {
             // 验证订单的状态
             Integer orderType = order.getOrderType();
             Integer orderStatus = order.getOrderStatus();
-            if (orderType != 0 && orderType != 1&& orderType != 3) {
+            if (orderType.intValue() != 0 && orderType.intValue() != 1&& orderType.intValue() != 3) {
                 throw new BusinessException("订单类型不正确,当前订单类型为:" + orderType);
             }
             if (orderStatus != 3) {
